% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoletFeatures.R
\name{computeFeatures.geoLet}
\alias{computeFeatures.geoLet}
\title{computeFeatures.geoLet lavora su piu}
\usage{
computeFeatures.geoLet(obj.geoLet, ROIName, feature.family = c("stat",
  "morph", "glcm", "rlm", "szm", "fractal"), filterPipeline = c(), px = "",
  py = "", error = 0.1)
}
\description{
Intiate an object of the class \code{geoLet}.This represents just the classname (f.extractor.sing.par)
}
