% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{DR.Bentzen}
\alias{DR.Bentzen}
\title{Function that calculates TCP according Bentzen (log Poisson) model}
\usage{
DR.Bentzen(doses, TD50 = 45, gamma50 = 1.5, a = 1)
}
\arguments{
\item{doses}{Either a \code{dvhmatrix} class object or a vector with nominal doses}

\item{TD50}{The value of dose that gives the 50\% of probability of outcome}

\item{gamma50}{The slope of dose/response curve at 50\% of probability}

\item{a}{Value for parallel-serial correlation in radiobiological response}
}
\description{
This function calculates the Tumor Control Probability according the
Bentzen (log Poisson) model.
}
\details{
In Bentzen paper dose/response curves were described both for tumor and normal tissue response.
In \pkg{moddicom} implementation \emph{log Poisson} model starts from the \code{\link{DR.Warkentin}} equation but 
calculating the relationship of response with the \emph{log} Dose as follows:
\deqn{TCP=\left (\frac{1}{2}  \right )^{\left (\frac{TD_{50}}{D}  \right )^{\frac{2\gamma_{50}}{log2}}}}
In the model equation \eqn{D} can be either the nominal dose or the \eqn{EUD} as calculated by \code{\link{DVH.eud}} function.
}
\references{
Bentzen SM, Tucker SL. \emph{Quantifying the position and steepness of radiation dose-response curves}. Int J Radiat Biol. 1997 May;71(5):531-42. Review. PubMed PMID: 9191898.
}
