% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvh_mod.R
\name{DR.fit.DoseVolume}
\alias{DR.fit.DoseVolume}
\title{Function for fitting Vdose and Dvolume to a given \code{dvhmatrix} object and outcome vector}
\usage{
DR.fit.DoseVolume(dvh, outcome, model, type = c("Vdose", "Dvolume"),
  CI = FALSE, CI.width = 0.95, epsilon = 1e-06)
}
\arguments{
\item{dvh}{A \code{dvhmatrix} object}

\item{outcome}{A vector of binary values (if not it will coerced to binary values) showing the observed outcome}

\item{model}{The model that will be used to fit the dose-volume-response data (see examples). Models with available 
the \code{\link[stats]{logLik}} method can be used belonging to classes \code{lm}, \code{glm} and \code{coxph}}

\item{type}{A character value representing the two type of dose-volume fitting as \code{Vdose} or \code{Dvolume}}

\item{CI}{Logcal Value for calculating the confidence interval of the dose-volume fitting parameter and for the \eqn{\alpha\beta} if modLQ is optioned \code{TRUE}}

\item{CI.width}{Width of confidence interval}

\item{epsilon}{Error limit for calculating the Log Likelihood function in determining the confidence interval of dose-volume parameter}
}
\description{
This function can be used to find the best \emph{Vdose} or \emph{Dvolume} values that describe
a given clinically observed outcome
}
\details{
The \emph{Vdose} is the value of the volume of a structure in a dose volume histogram that receives
a given level of dose. It is a widely used indicator and predictor of possible outcome when DVHs
have to be clinically evaluated. There are many papers that exploit the correct use of \emph{Vdose} for
clinical evaluation of treatment plans. The \code{DR.fit.DoseVolume} function used inside the \pkg{moddicom} package
allows to fit clinical data with \code{dvhmatrix} objects in order to detect the dose-volume relationship achievable
by dosimetric data. The dose-volume-outcome fitting is performed by iterative search, so, as far as the author know,
this is the first example of dose-volume-outcome fitting function that allows to find out \emph{the best fitting value}
of \emph{Vdose} or \emph{Dvolume}, rather than using an empyrical search from pre-fetermined values as usually reported in literature.
When setting the \code{model} parameter the users have to chose between a \code{\link[stats]{glm}} model and a \code{\link[survival]{coxph}} model
by writing the formula of the model \emph{without} the \code{dvhmatrix} object
}
\examples{
## NOT RUN DR.fit.DoseVolume(dvh = D, outcome = Dout, model = glm(formula = Dout ~ 1 , family = binomial(link = 'logit')), type = 'Vdose')
}
\references{
\emph{Special Considerations Regarding Absorbed-Dose and Dose-Volume Prescribing and Reporting in IMRT}. J ICRU. 2010 Apr;10(1):27-40. doi: 10.1093/jicru/ndq008. PubMed PMID: 24173325.

Graham MV, Purdy JA, Emami B, Harms W, Bosch W, Lockett MA, Perez CA. \emph{Clinical dose-volume histogram analysis for pneumonitis after 3D treatment for non-small cell lung cancer (NSCLC)}. Int J Radiat Oncol Biol Phys. 1999 Sep 1;45(2):323-9. PubMed PMID: 10487552.
}
