% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{DR.fit}
\alias{DR.fit}
\title{Fit a Dose/Response function}
\usage{
DR.fit(doses, outcome, DR.fun = c("Lyman", "Niemierko", "Bentzen", "Goitein",
  "Munro", "Okunieff", "Warkentin"), type = c("NTCP", "TCP"), CI = TRUE,
  C.I.width = 0.95, epsilon = 1e-06)
}
\arguments{
\item{doses}{Either a \code{dvhmatrix} class object or a vector with nominal doses}

\item{outcome}{A numeric vector of cases showing (1) and not showing (0) the outcome}

\item{DR.fun}{Dose/Response function, a character vector containing the name of one of the function in the package \pkg{moddicom}:
\code{Lyman}, \code{Niemierko}, \code{Bentzen}, \code{Goitein}, \code{Munro}, \code{Okunieff}, \code{Warkentin}.}

\item{type}{Function type: \code{NTCP}, Normal Tissue Complication Probability, or \code{TCP}, Tumor Control Probability}

\item{CI}{If \code{TRUE} it returns the value of confidence interval calulated by profile likelihood method}

\item{epsilon}{Lower interval for bisecting the MLE}

\item{CI.width}{The value of width of confidence interval to be returned if \code{CI = TRUE}}
}
\description{
This function fits an object of class \code{dvhatrix} or \code{numeric} to a given series of
outcome.
}
\details{
This function is a wrapper for fitting dose/response functions to a given set of binary outcomes. Currently
outcomes must be referred as a numeric vector  representing cases "showing the outcome" with the number 1, 
or "not showing the outcome" with the number 0. Model fitting is performed using Maximum Likelihood Estimation method,
implemented by wrapping R MLE functions set for specific Log Likelihood functions related to dose/response formulas.
If \code{doses} is a vector of nominal doses its length must be equal to the length of \code{outcome}, if \code{doses} is
a \code{dvhmatrix} class object the number of histograms must be equal to the length of \code{outcome}.
}
