% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{DR.generate}
\alias{DR.generate}
\title{Function for generating a series of outcomes related to a \code{dvhmatrix} class object}
\usage{
DR.generate(doses, a = 1, TD50 = 45, gamma50 = 1.5, DR.fun = c("Lyman",
  "Niemierko", "Bentzen", "Goitein", "Munro", "Okunieff", "Warkentin"))
}
\arguments{
\item{doses}{a \code{dvhmatrix} class object or a vector of nominal doses}

\item{a}{factor for calculating EUD value according \code{\link{DVH.eud}} function}

\item{TD50}{Dose that gives the 50\% probability of outcome}

\item{gamma50}{Slope of dose-response curve at TD50 of administered dose}

\item{DR.fun}{Dose/Response function, a character vector containing the name of one of the function in the package \pkg{moddicom}:
\code{Lyman}, \code{Niemierko}, \code{Bentzen}, \code{Goitein}, \code{Munro}, \code{Okunieff}, \code{Warkentin}.}
}
\value{
A vector of binary events (1 or 0).
}
\description{
This function can be used for simulating outcome related to a given vector of doses or
a \code{dvhmatrix} class object after setting the outcome related parameters \eqn{TD_{50}}{TD50}, \eqn{\gamma_{50}}{\gamma 50}
and \eqn{a}. If \code{doses} isn't a \code{dvhmatrix} class object and it is a vector of nominal doses
the \code{a} parameter is ignored.
}
