% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{DR.Goitein}
\alias{DR.Goitein}
\title{Function that calculates NTCP according Goitein (Bentzen) model}
\usage{
DR.Goitein(doses, TD50 = 45, gamma50 = 1.5, a = 1)
}
\arguments{
\item{doses}{Either a \code{dvhmatrix} class object or a vector with nominal doses}

\item{TD50}{The value of dose that gives the 50\% of probability of outcome}

\item{gamma50}{The slope of dose/response curve at 50\% of probability}

\item{a}{Value for parallel-serial correlation in radiobiological response}
}
\value{
A vector with NTCP(s) calculated according Goitein model.
}
\description{
This function calculates the Normal Tissue Complication Probability according the
Goitein (Bentzen) model.
}
\details{
This model is similar to Lyman model but it is function of \emph{logarithm} of the Dose.
Starting from the implementation of Lyman model: \deqn{NTCP=1/\sqrt{2\pi}\int_{-\infty}^{t}exp(-x^2/2)dx}
where: \deqn{v=\frac{V}{V_{ref}}} \deqn{t=(log(D)-TD_{50}(v))/(m*TD_{50}(v))} \deqn{TD(v)=TD(1)*v-n} In previous equations \eqn{V} is the
irradiated fraction volume and \eqn{V_{ref}} is the referenced volume for the given outcome, \eqn{m} is another way
to describe the slope of dose-response curve (see following lines).
In the \pkg{moddicom} implementation the parameters to be set in the model are \eqn{TD_{50}} and \eqn{\gamma_{50}}.
The model has been coded by adapting the original formula using these parameters in this way:
\deqn{t=(log(EUD)-TD_{50})/(m*TD_{50})} with:
\deqn{m=\frac{1}{\gamma_{50}\sqrt{2\pi}}} 
\eqn{D} can be either the nominal dose or the \eqn{EUD} as calculated by \code{\link{DVH.eud}} function.
}
\references{
Shipley WU, Tepper JE, Prout GR Jr, Verhey LJ, Mendiondo OA, Goitein M, Koehler AM, Suit HD. \emph{Proton radiation as boost therapy for localized prostatic carcinoma}. JAMA. 1979 May 4;241(18):1912-5. PubMed PMID: 107338.

Bentzen SM, Tucker SL. \emph{Quantifying the position and steepness of radiation dose-response curves}. Int J Radiat Biol. 1997 May;71(5):531-42. Review. PubMed PMID: 9191898.
}
