% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{DR.Lyman}
\alias{DR.Lyman}
\title{Function that calculates NTCP according Lyman/Kutcher/Burman model}
\usage{
DR.Lyman(doses, TD50 = 45, gamma50 = 1.5, a = 1)
}
\arguments{
\item{doses}{Either a \code{dvhmatrix} class object or a vector with nominal doses}

\item{TD50}{The value of dose that gives the 50\% of probability of outcome}

\item{gamma50}{The slope of dose/response curve at 50\% of probability}

\item{a}{Value for parallel-serial correlation in radiobiological response}
}
\value{
A vector with NTCP(s) calculated according LKB model.
}
\description{
This function calculates the Normal Tissue Complication Probability according the
Lyman/Kutcher/Burman model.
}
\details{
This model is a reformulation of probit model by using other parameters than mean and standard deviation.
In its original formulation the NTCP is given by the following equation: \deqn{NTCP=1/\sqrt{2\pi}\int_{-\infty}^{t}exp(-x^2/2)dx}
where: \deqn{v=\frac{V}{V_{ref}}} \deqn{t=(D-TD_{50}(v))/(m*TD_{50}(v))} \deqn{TD(v)=TD(1)*v-n} In previous equations \eqn{V} is the
irradiated fraction volume and \eqn{V_{ref}} is the referenced volume for the given outcome, \eqn{m} is another way
to describe the slope of dose-response curve (see following lines).
In the \pkg{moddicom} implementation the parameters to be set in the model are \eqn{TD_{50}} and \eqn{\gamma_{50}}.
The model has been coded by adapting the original formula using these parameters in this way:
\deqn{t=(EUD-TD_{50})/(m*TD_{50})} with:
\deqn{m=\frac{1}{\gamma_{50}\sqrt{2\pi}}} The relationship between Dose and Volume has been achieved by using the
equivalent uniform dose as calculated by \code{\link{DVH.eud}} function.
}
\references{
Burman C, Kutcher GJ, Emami B, Goitein M. \emph{Fitting of normal tissue tolerance data to an analytic function}. Int J Radiat Oncol Biol Phys. 1991 May 15;21(1):123-35. PubMed PMID: 2032883.
}
