% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{DR.Munro}
\alias{DR.Munro}
\title{Function that calculates TCP according Munro/Gilbert/Kallman model}
\usage{
DR.Munro(doses, TD50 = 45, gamma50 = 1.5, a = 1)
}
\arguments{
\item{doses}{Either a \code{dvhmatrix} class object or a vector with nominal doses}

\item{TD50}{The value of dose that gives the 50\% of probability of outcome}

\item{gamma50}{The slope of dose/response curve at 50\% of probability}

\item{a}{Value for parallel-serial correlation in radiobiological response}
}
\value{
A vector with TCP calculated according Munro/Gilbert/Kallman model.
}
\description{
This function calculates the Tumor Control Probability according the
Munro/Gilbert/Kallman model.
}
\details{
This model is an empyrical dose/response curve that fits experimental data. In their
paper authors assume this curve to be equivalent to a Poisson model. The original model equation is:
\deqn{TCP=e^{-EN_{0}e^{\frac{-D}{D_{0}}}}}
\eqn{E} is a numerical parameter that is related to tumor radiosensitivity, \eqn{N_{0}} is the total initial number of
tumor clonogenic cells, \eqn{D} is the delivered dose and \eqn{D_{0}} is the increment of dose that lowers survival 
to 37 per cent. In our implementation Munro/Gilbert/Kallman model has been referenced to \eqn{TD_{50}} and \eqn{\gamma_{50}} as follows:
\deqn{TCP=2^{e^{e\gamma_{50}(1-\frac{D}{TD_{50}})}}}
In the model equation \eqn{D} can be either the nominal dose or the \eqn{EUD} as calculated by \code{\link{DVH.eud}} function.
}
\references{
Munro TR, Gilbert CW. \emph{The relation between tumour lethal doses and the radiosensitivity of tumour cells}. Br J Radiol. 1961 Apr;34:246-51. PubMed PMID: 13726846.
}
