% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{DR.Niemierko}
\alias{DR.Niemierko}
\title{Function that calculates (N)TCP according Niemierko model}
\usage{
DR.Niemierko(doses, TD50 = 45, gamma50 = 1.5, a = 1)
}
\arguments{
\item{doses}{Either a \code{dvhmatrix} class object or a vector with nominal doses}

\item{TD50}{The value of dose that gives the 50\% of probability of outcome}

\item{gamma50}{The slope of dose/response curve at 50\% of probability}

\item{a}{Value for parallel-serial correlation in radiobiological response}
}
\value{
A vector with NTCP(s) calculated according Niemierko model.
}
\description{
This function calculates the Normal Tissue Complication Probability according the
Niemierko model.
}
\details{
This model can be used to compute Tumor Control Probability (TCP) too.
It is the translation of a \emph{loglogit} generalized linear model as function of \eqn{TD_{50}} and \eqn{\gamma_{50}}.
The model equation is: \deqn{(N)TCP=\frac{1}{1+ ({\frac{TD_{50}}{D}})^{4\gamma_{50}}}}
\eqn{D} can be either the nominal dose or the \eqn{EUD} as calculated by \code{\link{DVH.eud}} function.
}
\references{
Gay HA, Niemierko A. \emph{A free program for calculating EUD-based NTCP and TCP in external beam radiotherapy}. Phys Med. 2007 Dec;23(3-4):115-25. Epub 2007 Sep 7. PubMed PMID: 17825595.
}
