% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{DR.Warkentin}
\alias{DR.Warkentin}
\title{Function that calculates TCP according Warkentin (Poisson) model}
\usage{
DR.Warkentin(doses, TD50 = 45, gamma50 = 1.5, a = 1)
}
\arguments{
\item{doses}{Either a \code{dvhmatrix} class object or a vector with nominal doses}

\item{TD50}{The value of dose that gives the 50\% of probability of outcome}

\item{gamma50}{The slope of dose/response curve at 50\% of probability}

\item{a}{Value for parallel-serial correlation in radiobiological response}
}
\value{
A vector with TCP calculated according Warkentin (Poisson) model.
}
\description{
This function calculates the Tumor Control Probability according the
Warkentin model.
}
\details{
This model is the equivalent of the \emph{Poisson} model where the covariates and their coefficients
have been reported as function of \eqn{TD_{50}} and \eqn{\gamma_{50}}. In Warkentin paper model computation starts with formula:
\deqn{TCP=e^{-Np_{s}(D)}}
where \eqn{N} is the initial number of clonogens, \eqn{p_{s}(D)} is the survival fraction after the dose \eqn{D}.
The previous equation can be rewritten as function of \eqn{TD_{50}} and \eqn{\gamma_{50}}:
\deqn{TCP=\left (\frac{1}{2}  \right )^{e^{\left [2\gamma_{50}(1-D/D_{50})/ln2  \right ]}}}
Using the assumption of independent subvolumes, for the case of heterogeneous irradiation, 
the overall probability of tumor control is the product of the probabilities of killing all 
clonogens in each tumor subvolume described by the differential DVH:
\deqn{TCP=\prod_{i}TCP(D_{i},v_{i})}
Thus, for a given differential DVH \eqn{\left \{ D_{i},v_{i} \right \}}, the TCP can be calculated using the following two-parameter TCP formula:
\deqn{TCP=\left (\frac{1}{2}  \right )^{\sum_{i}v_{i}e^{\left [2\gamma_{50}(1-D/D_{50})/ln2  \right ]}}}
In the model equation \eqn{D} can be either the nominal dose or the \eqn{EUD} as calculated by \code{\link{DVH.eud}} function.
}
\references{
Warkentin B, Stavrev P, Stavreva N, Field C, Fallone BG. \emph{A TCP-NTCP estimation module using DVHs and known radiobiological models and parameter sets}. J Appl Clin Med Phys. 2004 Winter;5(1):50-63. Epub 2004 Jan 1. PubMed PMID: 15753933.
}
