% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvh_mod.R
\name{DVH.absolute}
\alias{DVH.absolute}
\title{Converts relative \code{dvhmatrix} class objects to absolute}
\usage{
DVH.absolute(dvh)
}
\arguments{
\item{dvh}{A \code{dvhmatrix} class object}
}
\value{
A \code{dvhmatrix} class object.
}
\description{
Function that converts an object of \code{dvhmatrix} class where DVH are stored in relative mode into
another \code{dvhmatrix} class object where DVH are stored in absolute mode.
}
\examples{
## generate a dataset of relative DVHs
a<-DVH.generate(dvh.number = 10, vol.distr = "relative")
b<-DVH.absolute(dvh = a)
}
