% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvh_mod.R
\name{DVH.baseStat}
\alias{DVH.baseStat}
\title{Function for calculating the mean and median dvh with related confidence intervals}
\usage{
DVH.baseStat(dvh, C.I.width = 0.95, n.boot = 2000)
}
\arguments{
\item{dvh}{A \code{dvhmatrix} class object}

\item{C.I.width}{The width of confidence interval}

\item{n.boot}{The number of bootstrapped dvhs for computing the quantile in C.I. calculation}
}
\value{
A \code{dvhmatrix} object where 7 columns: Dose, mean DVH, low C.I. of mean DVH, high C.I. of mean DVH, 
        median DVH, low C.I. of median DVH, high C.I. of median DVH.
}
\description{
This function calculates the mean dvh from a \code{dvhmatrix} class object. The mean dvh is
             calculated with its confidence interval that is given by a bootstrapped dvh series from
             the dvh given in the \code{dvh} object.
}
\examples{
## creates a dvhmatrix class object with 100 DVHs
a<-DVH.generate(dvh.number = 100, dvh.type = "cumulative", vol.distr = "relative")
b<-DVH.baseStat(dvh = a)
## plot the dvhmatrix object "b", showing both mean, 
## median DVH and corresponding confidence intarvals
plot(b)
}
