% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvh_mod.R
\name{DVH.cum.to.diff}
\alias{DVH.cum.to.diff}
\title{Function that converts cumulative DVHs into differential ones}
\usage{
DVH.cum.to.diff(dvh)
}
\arguments{
\item{dvh}{Either an object of class \code{dvhmatrix} or \code{matrix} type.}
}
\value{
Either an object of class \code{dvhmatrix} or a \code{matrix} according the argument \code{dvh}.
}
\description{
Function that converts an object of class \code{dvhmatrix} or a simple \code{matrix} that 
             represents a cumulative DVH into a differential one.
}
