% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvh_mod.R
\name{DVH.extract}
\alias{DVH.extract}
\title{Extracts a DVH from a vector of dose bins}
\usage{
DVH.extract(x, max.dose = NULL, dose.bin = 0.25,
  dvh.type = c("differential", "cumulative"), vol.distr = c("relative",
  "absolute"), createObj = TRUE, volbin.side = NULL, voxel.volume = NULL,
  total.volume = NULL)
}
\arguments{
\item{x}{vector of dose bins.}

\item{max.dose}{Upper dose bound for limiting the DVH computation.}

\item{dose.bin}{The dose bin for DVH computation in Gy.}

\item{dvh.type}{The type of volume distribution to be created: \code{differential} or \code{cumulative}.}

\item{vol.distr}{Defines if the volume bins have to be divided by the total volume of the structure (\code{relative}) or not (\code{absolute}).}

\item{createObj}{if \code{TRUE} returns a \code{dvhmatrix} class object.}

\item{volbin.side}{The value of the side of each cube (in cm) that builds the final volume of the structure.}

\item{voxel.volume}{The value of the volume of a single voxel (in cc) to compute the ROI total volume}

\item{total.volume}{The value of the total volume of the ROI (in cc)}
}
\value{
Either an object of class \code{dvhmatrix} or a \code{matrix} according the argument \code{dvh} type.
}
\description{
Function that given a vector of dose bins (either a vector got from sampling a 3D mesh or a vector of
simple values of dose) extracts the DVH that summarizes that vector. Either \code{volbin.side}, \code{voxel.volume} or \code{total.volume} have to be
not \code{NULL} values. The priority list for computing the volume (if more than one argument is not \code{NULL}) is total.volume \code{->} voxel.volume \code{->} volbin.side.
}
\examples{
## simulate a vector of dose bins
doses <- c(rnorm(n = 10000, mean = 45, sd = 3), rnorm(n = 7000, mean = 65, sd = 2.5))

## creates a dvhmatrix class object
DVH<-DVH.extract(x = doses)
}
