% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvh_mod.R
\name{DVH.generate}
\alias{DVH.generate}
\title{Function that creates a \code{dvhmatrix} class object}
\usage{
DVH.generate(dvh.number, type = c("random", "convex", "concave", "mix"),
  dvh.type = c("differential", "cumulative"), vol.distr = c("relative",
  "absolute"), max.dose = 75, dose.bin = 0.5, volbin.side = 2.5,
  min.vol = 180, max.vol = 220)
}
\arguments{
\item{dvh.number}{The number of DVHs to be generated}

\item{type}{The type of DVH to be generated: \code{random} generates all possible variation in dose distribution, 
\code{convex} for DVH similar to PTV structures with good dose homogeneity to high levels,
\code{concave} for DVH with low level of dose delivered to the volume as a spared structure,
\code{mix} for DVH where dose distribution is averaged in the whole volume spectrum.}

\item{dvh.type}{The type of volume distribution to be created: \code{differential} or \code{cumulative}.}

\item{vol.distr}{Defines if the volume bins have to be divided by the total volume of the structure (\code{relative}) or not (\code{absolute}).}

\item{max.dose}{The upper bound of the dose distribution to be simulated.}

\item{dose.bin}{The dose bin in Gy to compute the value of volume parts in the final DVH.}

\item{volbin.side}{The value of the side of each cube (in mm) that builds the final volume of the simulated structure.}

\item{min.vol}{The minimum volume of the range to be simulated.}

\item{max.vol}{The maximum volume of the range to be simulated.}
}
\value{
An object of \code{dvhmatrix} class.
}
\description{
Creates an object of class \code{dvhmatrix}.
}
\details{
Dose Volume Histograms (DVH) are the basic representation of how the radiation doses distribute inside structures.
         Usually they are shown in two forms: \code{differential} or \code{cumulative} being related each other because
         \code{cumultive} DVHs are the integral form of \code{"differential"} ones. This function provides a siumlated series
         of structures with the features defined by function parameters. The simulated series are useful for producing simulated
         Dose/Response models using the modeling functions defined in this package.
         class objects.
}
\examples{
## creates a dvhmatrix object with 200 differential - relative histograms
a<-DVH.generate(dvh.number=200, dvh.type="differential", vol.distr="relative")

## creates a dvhmatrix object containing 150 cumulative - absolute histograms
## maximum dose in simulated series is 60 Gy
b<-DVH.generate(dvh.number=150, dvh.type="cumulative", vol.distr="absolute", max.dose=60)
}
\references{
Van den Heuvela F. \emph{Decomposition analysis of differential dose volume histograms.} Med Phys. 2006 Feb;33(2):297-307. PubMed PMID: 16532934.
}
