% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvh_mod.R
\name{DVH.lq.correct}
\alias{DVH.lq.correct}
\title{Function that returns the linear quadratic correction of dose bin in a \code{dvhmatrix} object}
\usage{
DVH.lq.correct(dvh, ref.frac = 2, nf, alphabeta = 3)
}
\arguments{
\item{dvh}{A \code{dvhmatrix} object}

\item{ref.frac}{Reference fractionation, default value 2}

\item{nf}{Fractions number for the input dvh}

\item{alphabeta}{\eqn{\alpha\beta} ratio for computing LQ cvonversion, default value 3}
}
\value{
A \code{dvhmatrix} object with converted dose bins
}
\description{
This function appies the linear-quadratic correction for the dose in a \code{dvmatrix} object by 
converting each dose bins as function of a given \eqn{\alpha\beta} ratio (default value is 3, valid for late outcome).
}
