% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvh_mod.R
\name{DVH.merge}
\alias{DVH.merge}
\title{Merge two different \code{dvhmatrix} class objects into one}
\usage{
DVH.merge(receiver = NULL, addendum = NULL)
}
\arguments{
\item{receiver}{The \code{dvhmatrix} object that will receive the \code{addendum} object.}

\item{addendum}{The \code{dvhmatrix} object to be merged with \code{addendum}.}
}
\value{
A \code{dvhmatrix} class object.
}
\description{
This function can e used to create a \code{dvhmatrix} class object from two different
             ones. The slots of the resulting object will be the same in the \code{receiver} one,
             so some convertions can be automatically realized to create the homogeneous final result.
}
\examples{
## creates two different dvhmatrx objects
a<-DVH.generate(dvh.number = 100, dvh.type="differential", vol.distr = "relative")
b<-DVH.generate(dvh.number = 100, dvh.type="cumulative", vol.distr = "absolute")
ab<-DVH.merge(receiver = a, addendum = b)
}
