% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvh_mod.R
\name{DVH.relative}
\alias{DVH.relative}
\title{Converts absolute \code{dvhmatrix} class objects to relative}
\usage{
DVH.relative(dvh)
}
\arguments{
\item{dvh}{A \code{dvhmatrix} class object}
}
\value{
A \code{dvhmatrix} class object.
}
\description{
Function that converts an object of \code{dvhmatrix} class where DVH are stored in absoulte mode into
another \code{dvhmatrix} class object where DVH are stored in relative mode.
}
\examples{
## generate a dataset of absolute DVHs
a<-DVH.generate(dvh.number = 10, vol.distr = "absolute")
b<-DVH.relative(dvh = a)
}
