% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvh_mod.R
\name{DVH.Vdose}
\alias{DVH.Vdose}
\title{Function for extracting the V-Dose from cumulative DVH(s)}
\usage{
DVH.Vdose(dvh, Dose)
}
\arguments{
\item{dvh}{A \code{dvhmatrix} class object}

\item{Dose}{The dose for calculating the V-Dose value(s)}
}
\value{
A vector containing the value(s) of V-Dose(s).
}
\description{
Function for calculating the value of V-Dose of a \code{dvhmatrix} object
}
\details{
V-Dose is, in a given Dose Volume Histogram, the value of the volume of the structure receiving
         \strong{at least} the chosen level of dose.
}
\examples{
## create a dvhmatrix class object
a<-DVH.generate(dvh.number = 100)
DVH.Vdose(dvh = a, Dose = 50)
}
