% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{dvhmatrix-class}
\alias{dvhmatrix-class}
\title{dvmmatrix S4 class}
\description{
This class handles the matrices containing the DVH(s)
}
\section{Slots}{

\describe{
\item{\code{dvh}}{The numeric matrix containing the DVH(s). The 1st column represents the dose steps, the other(s) represent(s) the volume values.}

\item{\code{dvh.type}}{A character slot giving the \code{"cumulative"} or \code{"differential"} representation of the volume(s).}

\item{\code{vol.distr}}{A character slot setting the \code{"absolute"} or \code{"relative"} value of the volume(s).}

\item{\code{volume}}{A numeric vector giving the value(s) of the total volume(s) of the structure(s) in the DVH(s).}
}}

