% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoletFeatures.R
\name{f.extractor.pluri.LoG.par}
\alias{f.extractor.pluri.LoG.par}
\title{f.extractor.pluri.LoG.par}
\usage{
f.extractor.pluri.LoG.par(path, ROIName, feature.family = c("stat", "morph",
  "glcm", "rlm", "szm", "fractal"), interpolate = FALSE, px = "", py = "",
  error = 0.1, fileName = "tmp.f.extractor.pluri.par.RData",
  forceRecalculus = TRUE, from.sigma = 1, to.sigma = 2, def.by = 0.1,
  strategy = "fixed", sigma.array = c())
}
\description{
Instantiate an object of the class \code{geoLet}.This represents just the classname (f.extractor.pluri.LoG.par)
}
