% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoletFeatures.R
\name{f.extractor.sing.par}
\alias{f.extractor.sing.par}
\title{f.extractor.sing.par}
\usage{
f.extractor.sing.par(path, ROIName, feature.family = c("stat", "morph",
  "glcm", "rlm", "szm", "fractal"), filterPipeline = list(),
  interpolate = FALSE, px = "", py = "", error = 0.001,
  fileName = "tmp.f.extractor.sing.par.RData", forceRecalculus = TRUE)
}
\arguments{
\item{path}{The path to be browsed for folders containing DICOM images and Structure Set of patients}

\item{ROIName}{A \code{character} object containing the name(s) of ROIs to be analyzed. If there is more than one ROI (multiple references) or no ROI is 
listed the analysis is stopped}

\item{feature.family}{The list of features to be extracted, by default all}

\item{filterPipeline}{A \code{list} containint the filter pipeline to be applied to images before feature extraction}

\item{interpolate}{A \code{logical} value, by default \code{FALSE}, for interpolating sigma values in filters using sigma for computation as LoG}

\item{px}{The pixel spacing in \emph{x} direction for interpolating sigma}

\item{py}{The pixel spacing in \emph{y} direction for interpolating sigma}

\item{error}{A \code{numeric} value setting the threshold for applying the interpolation}

\item{forceRecalculus}{A \code{logical} value, by defaul \code{TRUE}, for forcing recalulus}
}
\description{
Instantiate an object of the class \code{geoLet}. This represents just the classname (f.extractor.sing.par)
}
\details{
Using an object of class \code{ROImap} it is possible to list all the ROIs contained in an explored directory for the analysis. It is possible to analyze
different ROIs names referencing to the same structure in the different studies by putting al their names a a vector of \code{character} objects in the argument \code{ROIName}
}
