% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter.R
\name{FIL.applyFilter}
\alias{FIL.applyFilter}
\title{Apply a filter to a 2d or a 3d array wow}
\usage{
FIL.applyFilter(arr2App, kernel.type, sigma = 1.4)
}
\arguments{
\item{arr2App}{is the array containing the image that should be filtered}

\item{kernel.type}{is the kernel you want to use: \code{gaussian}, \code{laplacian}, \code{LoG},\code{shrpen}, \code{emboss}, \code{sobel}}

\item{sigma}{is the \code{sigma} value for gaussian filter}
}
\value{
the filteres array (same geometry of the one in input)
}
\description{
It applies filter to a 2d or 3d array. '...' is banned, so shit cannot flies ( ;) ).
}
\examples{
\dontrun{
# Create an instante of new.mmButo and load some cases
a<-array( 0, dim=c(10,10,10))
a[5,5,5]<-177; a[5,6,5]<-198; a[6,6,5]<-45; a[6,5,5]<-12;
b<-FIL.applyFilter( a , kernel.type="gaussian", sigma=1.4)
}#' 
}
