% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter.R
\name{FIL.applyFilterToStudy}
\alias{FIL.applyFilterToStudy}
\title{DEPRECATED!!!!:  Apply a filter to mmButo object}
\usage{
FIL.applyFilterToStudy(obj.mmButo, ROINameForNormalization = NA,
  valueForNormalization = NA, ROIName, filter.pipeline,
  collection = "default", cropResult = TRUE, scaleFactor = "voxel")
}
\arguments{
\item{obj.mmButo}{is the \code{new.mmButo} object}

\item{ROINameForNormalization}{the name of the ROI used to normalize the signal}

\item{valueForNormalization}{the highest value used to normalize the 'ROINameForNormalization' voxel values}

\item{ROIName}{the name of the ROI to extract}

\item{filter.pipeline}{is a \code{list} where is indicated the pipeline of filtering to be applied (in sequence)}

\item{collection}{is the collection: the default is '\code{default}'}

\item{cropResult}{is a boolean (\code{TRUE} or \code{FALSE}) which indicates if the result should be cropped or not, in order to save memory. Default is \code{TRUE}}

\item{scaleFactor}{can be 'voxel' or 'space'. If 'voxel' (the default) it consider the sam sigma values (if specified) for all the geoLet object, if 'space' it normalize the sigma according to the different pixelSpacing.}
}
\value{
a list containing the filtered images (cropped or not)
}
\description{
It applies filter to an mButo object to simplify RADIOMICS issues
}
\examples{
\dontrun{
# DEPRECATED!!!!!
} 
}
