% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter.R
\name{FIL.getFilteredVoxelCube}
\alias{FIL.getFilteredVoxelCube}
\title{Return a filtered VoxelCube}
\usage{
FIL.getFilteredVoxelCube(obj.geoLet, ROIName, kind.of.filter, sigma = 0.8)
}
\value{
an list of items
}
\description{
returns a filtered voxelcube (WIP)
}
