% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoLet.R
\name{geoLet}
\alias{geoLet}
\title{class for loading and presenting DICOM data}
\usage{
geoLet(ROIVoxelMemoryCache = TRUE, folderCleanUp = FALSE,
  loadXMLInCache = FALSE, loadRAWInCache = TRUE,
  defaultExtension = "*.dcm")
}
\description{
Instantiate an object of the class \code{geoLet}.This represents just the classname, 
              methods are exposed with the technique of 'closure'.
              In order to see manuals for the single mathods, consider the vignette or use the 
              available for the following wrapping functions:
              \itemize{
              \item \code{GLT.openDICOMFolder( );} : to load a DICOM series into an geoLet object
              \item \code{GLT.getImageVoxelCube( );} : to get the ImageVoxelCube stored into a geoLet object
              \item \code{GLT.getPixelSpacing( );} : to get the pixelSpacing (x,y,z) of the main ImageVoxelCube stored into a geoLet object
              \item \code{GLT.getROIList( );} : to get the list of the ROI defined in a geoLet object
              \item \code{GLT.getTag( );} : to get a single DICOM-tag of a DICOM file loaded into a geoLet object
              \item \code{GLT.getROIVoxels( );} : to get the IMAGE Voxels geometrically located into a ROI, for a given geoLet object
              \item \code{GLT.extractDoseVoxels( );} : to get the DOSE Voxels geometrically located into a ROI, for a given geoLet object
              \item \code{GLT.calculateDVH( );} : to get the DVH calculated from a geoLet object
              }
              The original methods for the class geoLet can also be invocked using the same name without the previx 'GTL.', i.e.:
}
\examples{
\dontrun{
# first of all create an object obj.1 and obj.2
obj.1<-geoLet()
obj.2<-geoLet()

# now load a DICOM serie using the methods for obj.1 and the wrapping function for obj.2
# The result is the same: the only difference is some froceries available in the second way (i.e.: help online)

obj.1$openDICOMFolder(pathToOpen='./DICOMSeries/pat001' );
obj.2<-GLT.openDICOMFolder(obj = obj.2, pathToOpen='./DICOMSeries/pat001' );#' 

}
}
