% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoLet.wrappingFunctions.R
\name{GLT.calculateDVHs}
\alias{GLT.calculateDVHs}
\title{Returns the DVH calculated from data stored into a geoLet object}
\usage{
GLT.calculateDVHs(obj.geoLet, ROIName, newPixelSpacing = NA,
  justTheDVH = TRUE, verbose = FALSE, forceReCalculus = FALSE,
  fastEngine = TRUE, decimation = FALSE, decimation.percentage = 0.8,
  smoothing = FALSE, smoothing.iterations = 10)
}
\arguments{
\item{obj.geoLet}{the object geoLet to load the DICOM serie in}

\item{ROIName}{the name of the ROI you are interested in having the DVH}

\item{newPixelSpacing}{Optional. By default it uses same pixelValues of the main CT scan but if you want to interpolate the space in order to have a bigger detail you can increase the precision. Pay attention: it can take a lot o memory!!!}

\item{justTheDVH}{Optional. \code{TRUE} by default. Because of this function is also able to calculate other ancillary data structures, if you want you can set this parameter to \code{FALSE} and get also the IMAGE voxel cube and te DOSE voxel cube (sampled at the same points in the 3D space)}

\item{verbose}{Optional. \code{FALSE} by default. Set to \code{TRUE} it allow to see some logs during the computation}

\item{forceReCalculus}{Optional. \code{FALSE} by default. Due to a caching system, implemented to improve performances, If a previous calculus has been interrupted the afterwards computations (on the same ROI)  could have problems. If a computation has been interrupted it is a good practice to set this parameter to \code{TRUE} in the next computation in order to let the algorithm to re-build anything.}

\item{fastEngine}{\code{TRUE} by default. By default it uses the \code{vcgClostKD()} function which should be quicker but probably a bit less accurate (?). If you want to try something different you can set this parameter on \code{FALSE} and it will use the most classical \code{vcgClost}}

\item{decimation}{Optional. \code{FALSE} by default. Set it to \code{TRUE} if you want to enable decimation to the mesh structure}

\item{smoothing}{Optional, \code{FALSE} by default. This parameter, if se to \code{TRUE} allow to smooth the mesh by a numerical factor indicated in the parameter \code{smoothing.iterations}.}

\item{decimationpercentage.}{Optional. 0.8 by default. If the parameter \code{decimation} is set to \code{TRUE} this parameter indicates the percentage of triangles that should be sacrified in the first (and biggest) mesh-model. Killing triangles allows to have an easier to handle mesh and less memory-consumer but reduce the quality of the approximation. In most cases the optimal value has to be defined empirically. If \code{decimation} is set to \code{FALSE} dont's waste your time in tuning this parameter: decimation will not be performed!}

\item{smoothing.terations}{Optional, 10by default. Setting this parameter makes sense only if \code{smoothing} was previously set to \code{TRUE}. Increasing the number of iterations we can improve the "smoothing" applied to the original mesh.}
}
\value{
Depending on the status of the parameter \code{justTheDVH}. If it is set to \code{TRUE} the function returns a \code{dvhmatrix} object. If set to \code{FALSE} it returns a list of three elements:
\itemize{
   \item{ \code{voxelCube.CT} this is a 3D array which contains the voxel cube of the original CT scan }
   \item{ \code{voxelCube.Dose} this is a 3D array which contains the voxel cube of the DOSES. The dimension is the same of the \code{voxelCube} because the dose value in the space has been interpolated (via triliear interpolation) in order to overlap the centroids of the voxels between \code{voxelCube.CT} and \code{voxelCube.Dose}. This should allow to quickly plot the expected dose value on a specific pixel/voxel in the CT scan. }
   \item{ \code{DVHobj}  a \code{dvhmatrix} object }    
}
}
\description{
This function calculates DVH starting from DICOM RT-DOSE objects, previously loaded into a geoLet object.

              N.B: if you are reading this help, your life is going to end in 10, 9, 8, 7, ...
}
\examples{
\dontrun{

obj<-geoLet()
obj$openDICOMFolder(pathToOpen='./DICOMSeries/pat001' );
a<-obj$GLT.calculateDVH( Structure = "GTV" )

}
}
