% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoLet.wrappingFunctions.R
\name{GLT.getImageVoxelCube}
\alias{GLT.getImageVoxelCube}
\title{Returns an IMAGE voxel cube}
\usage{
GLT.getImageVoxelCube(obj.geoLet, ps.x = NA, ps.y = NA, ps.z = NA)
}
\arguments{
\item{obj.geoLet}{the object geoLet to load the DICOM serie in;}

\item{ps.x}{The desided pixelspaxing along the x-axes. If not specifies, by default, it will keep the original pixel spacing. Pay attention: lowest pixel spacings can cause long computation time and crashes due to memory swapping on disk.}

\item{ps.y}{The desided pixelspaxing along the y-axes. If not specifies, by default, it will keep the original pixel spacing. Pay attention: lowest pixel spacings can cause long computation time and crashes due to memory swapping on disk.}

\item{ps.z}{come on, genius, try to guess....}
}
\value{
it returns a 3D array containing ALL the IMAGE voxel cubes. No bullshit.
}
\description{
Returns the primary IMAGE voxel cube stored into a geoLet object

              N.B: pay attention, it can cause infertility
}
\examples{
\dontrun{

obj<-geoLet()
obj$openDICOMFolder(pathToOpen='./DICOMSeries/pat001' );
imageVoxelCube<-GLT.getImageVoxelCube(obj = obj );

}
}
