% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoLet.wrappingFunctions.R
\name{GLT.getPixelSpacing}
\alias{GLT.getPixelSpacing}
\title{Returns the pixelSpacing along three axes of a geoLet DICOM image serie}
\usage{
GLT.getPixelSpacing(obj.geoLet, seriesInstanceUID = NA)
}
\arguments{
\item{obj.geoLet}{the object geoLet to load the DICOM serie in;}

\item{seriesInstanceUID}{the seriesInstance UID you are interested in. It can also be not specified: in this case it return the pixelSpacing of the 'main' image series (which is normally che interested one)}
}
\value{
Returns an array containing the pixelspacing along x, y and z
}
\description{
Returns the pixelSpacing of a desired DICOM Series previously stored in a geoLet Object

              N.B: pay attention, it can bite
}
\examples{
\dontrun{

obj<-geoLet()
obj$openDICOMFolder(pathToOpen='./DICOMSeries/pat001' );
pixelSpacingValue<-GLT.getPixelSpacing(obj = obj );

}
}
