% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoLet.wrappingFunctions.R
\name{GLT.getROIList}
\alias{GLT.getROIList}
\title{Returns the list of the ROI available in a geoLet object}
\usage{
GLT.getROIList(obj.geoLet)
}
\arguments{
\item{obj.geoLet}{the object geoLet to load the DICOM serie in;}
}
\value{
It returns a matrix where in the second row you can read the ROIName stored into the DICOM RT-struct. Please consider that not alwasy the ROIPointList are really associated to the shown ROIs. This depends from how the DICOM Serie has been exported from TPS.
}
\description{
Returns the list of the ROI previously loaded in a geoLet object

              N.B: pay attention, it can seduce your wife while you use it
}
\examples{
\dontrun{

obj<-geoLet()
obj$openDICOMFolder(pathToOpen='./DICOMSeries/pat001' );
obj$getROIList( obj.geoLet = obj )

}
}
