% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoLet.wrappingFunctions.R
\name{GLT.getTag}
\alias{GLT.getTag}
\title{Returns the value of a tag of a specified DICOM object}
\usage{
GLT.getTag(obj.geoLet, tag, fileName = "")
}
\arguments{
\item{obj.geoLet}{the object geoLet to load the DICOM serie in;}

\item{tag}{it is a string which indicates the DICOM tag you are interested in. I.e.: "0010,0010" for the PatientID, "0010,0020" for the Patient Name and so on. Please, refer to the DICOM standard for the full list of possible tags. Consider also that this function is designed to return just the most common tags: the most exhostic ones (for example the nested ones) cannot be returned or, anyway, the result is not sure.}

\item{fileName}{If you know the fileName of the DICOM object you are interested in, you can specify it in this parameter. This parameter is optional: if not specified geoLet will give back, by default, the tag taken from the first image of the main image serie}
}
\value{
It returns a matrix where in the second row you can read the ROIName stored into the DICOM RT-struct. Please consider that not alwasy the ROIPointList are really associated to the shown ROIs. This depends from how the DICOM Serie has been exported from TPS.
}
\description{
Returns the value of a tag of a specified DICOM object (image file, RTStruct, ...)

              N.B: pay attention, cachexia is one of the most common effect of this function
}
\examples{
\dontrun{

obj<-geoLet()
obj$openDICOMFolder(pathToOpen='./DICOMSeries/pat001' );
obj$GLT.getTag( obj.geoLet = obj, tag = "0010,0020")

}
}
