% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoLet.wrappingFunctions.R
\name{GLT.openDICOMFolder}
\alias{GLT.openDICOMFolder}
\title{Load a DICOM serie into a geoLet object}
\usage{
GLT.openDICOMFolder(obj.geoLet, pathToOpen, defaultExtension = "*.dcm")
}
\arguments{
\item{obj.geoLet}{the object geoLet to load the DICOM serie in;}

\item{pathToOpen}{the path where the DICOM serie is locate, on the filesystem}

\item{defaultExtension}{the 'typical' extension of your DICOM files. Default is '.dcm'}
}
\value{
wel.... nothing. It directly load the serie into the passed geoLet object. Technically it shouldn't work but apparently it does.
}
\description{
Allow to load a DICOM serie into a geoLet object.

              N.B: pay attention, it can explode if not properly used
}
\examples{
\dontrun{

obj<-geoLet()
GLT.openDICOMFolder(obj = obj, pathToOpen='./DICOMSeries/pat001' );

It is equivalent to:

obj<-geoLet();
obj$openDICOMFolder(pathToOpen='./DICOMSeries/pat001' );
}
}
