% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoLet.wrappingFunctions.R
\name{GLT.rotateToAlign}
\alias{GLT.rotateToAlign}
\title{Remap the ROI points according to the raster geomery of image}
\usage{
GLT.rotateToAlign(obj.geoLet, ROIName)
}
\arguments{
\item{obj.geoLet}{the object geoLet to load the DICOM serie in}

\item{ROIName}{the name of the ROI you are interested in remapping}
}
\value{
It return a list of points located in the bitmap geometry of the CT (or MRI) scan. By this function you can in a quite easy way identify ROI points on the image slices
}
\description{
Remap the 3D points of the chosen ROI in the 3D bitmap geometry of the main image

              N.B: if you are reading this help, Clostridium difficile will bless you...
}
\examples{
\dontrun{

obj<-geoLet()
obj$openDICOMFolder(pathToOpen='./DICOMSeries/pat001' );
a<-obj$GLT.calculateDVH( ROIName = "GTV" )

}
}
