% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoletFeatures.R
\name{map.ROI}
\alias{map.ROI}
\title{map.ROI}
\usage{
map.ROI(path)
}
\arguments{
\item{path}{A \code{character} value containing the path to browse for DICOM studies}
}
\value{
An object of class \code{ROImap} that is a list containing all mapped ROIs in a series of studies
}
\description{
Function for mapping all ROI names in a series a studies
}
\details{
DICOM RT structure set files often contain multiple \emph{ROI names} referencing to the same structure. \code{moddicom} is desinged to analyze only
one structure at a time. Using this function a \code{ROImap} class object is created, listing all ROI names for each case inside the borwsed directory.
Using the method \code{print.ROImap} a table showing all ROI names and their frequency in the explored path is shown.
}
\examples{
## NOT RUN
ListOfROI <- map.ROI(path = '/my_path')
## to get the table of ROIs in the explored path
print(ListOfROI)
}
