% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmButo.wrappingFunctions.R
\name{MBT.getAllPixelSpacing}
\alias{MBT.getAllPixelSpacing}
\title{Retrieve the pixelSpacing values of the stored series}
\usage{
MBT.getAllPixelSpacing(obj.mmButo)
}
\arguments{
\item{obj.mmButo}{the object mmButo}
}
\value{
a list of arrays; every array contains the x,y,z dimension of the voxel for the interested serie.
}
\description{
In quickly allows to read the pixel spacing of the series loaded in a mmButo object. Useful when you are in duobt of having differen pixelSpacing and you are not sure you can work without considering the spatial dimension of the voxels..

              N.B: pay attention.... Splat! Ok, bless you!
}
\examples{
\dontrun{

obj<-mmButo()
MBT.getROIVoxel(obj.mmButo = obj, ROIName='GTV' );

}
}
