% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmButo.wrappingFunctions.R
\name{MBT.getImageVoxel}
\alias{MBT.getImageVoxel}
\title{Retrieve the imageVoxelCube of the main images}
\usage{
MBT.getImageVoxel(obj.mmButo, ps.x = NA, ps.y = NA, ps.z = NA)
}
\arguments{
\item{obj.mmButo}{the object mmButo}

\item{ps.x}{Optional, default \code{NA}. If you want to interpolate the image to a new step on the x-axes you can specif here the new PixelSpaginc along the x direction}

\item{ps.y}{Optional, default \code{NA}. If you want to interpolate the image to a new step on the Y-axes you can specif here the new PixelSpaginc along the y direction}

\item{ps.z}{Optional, default \code{NA}. come on... try to imagine....}
}
\value{
a list of 3D arrays
}
\description{
It gives back the complete MRI/CT/PET scans, evantually interpolated to the wished step

              N.B: pay attention.... Splat! Ok, bless you!
}
\examples{
\dontrun{

obj<-mmButo()
MBT.getImageVoxel(obj.mmButo = obj);

}
}
