% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmButo.wrappingFunctions.R
\name{MBT.getROIVoxel}
\alias{MBT.getROIVoxel}
\title{Returns the lists of voxels internal to a given ROIName}
\usage{
MBT.getROIVoxel(obj.mmButo, ROIName)
}
\arguments{
\item{obj.mmButo}{the object mmButo}

\item{pathToOpen}{the path where the DICOM serie is locate, on the filesystem. It should point to a path where it can find A SET OF FOLDERS, ONE FOR EACH PATIENT}
}
\value{
wel.... nothing.
}
\description{
Once a mmButo object is loaded, with \code{loadCollection} method, it allows to extract all the voxels within a given ROINAME. The behaviour is similar to the \code{GLT.getImageROIVoxels} but is able to work on all the DICOM studies stored in the object.

              N.B: pay attention.... ops! Your wallet is empty!
}
\examples{
\dontrun{

obj<-mmButo()
MBT.getROIVoxel(obj.mmButo = obj, ROIName='GTV' );

}
}
