% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmButo.wrappingFunctions.R
\name{MBT.loadCollection}
\alias{MBT.loadCollection}
\title{Load a set of DICOM studies into a mmButo structure}
\usage{
MBT.loadCollection(obj.mmButo, pathToOpen)
}
\arguments{
\item{obj.mmButo}{the object mmButo}

\item{pathToOpen}{the path where the DICOM serie is locate, on the filesystem. It should point to a path where it can find A SET OF FOLDERS, ONE FOR EACH PATIENT}
}
\value{
wel.... nothing.
}
\description{
Allow to load a set of DICOM studies (but just with ONE seriesInstanceUID for each study) into an mmButo structure. 

              N.B: pay attention, using it you c n loo e the sign l
}
\examples{
\dontrun{

obj<-mmButo()
MBT.loadCollection(obj.mmButo = obj, pathToOpen='./DICOMSeries/pat001' );

}
}
