% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmButo.R
\name{mmButo}
\alias{mmButo}
\title{class for loading multiple sets of DICOM studies from filesystem}
\usage{
mmButo(folderCleanUp = FALSE, loadXMLInCache = FALSE,
  loadRAWInCache = TRUE)
}
\description{
Instantiate an object of the class \code{new.mmButo}. This represents just the classname,
              for each instantiated object many methods are available not i S3 or S4 but by closures method.
              The available methods are:
}
