% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvh_mod.R
\docType{methods}
\name{plot,DoseVolumeModel,ANY-method}
\alias{plot,DoseVolumeModel,ANY-method}
\title{method for plotting the Log-Likelihood values in \code{DoseVolumeModel} class object}
\usage{
\S4method{plot}{DoseVolumeModel,ANY}(x, ShowFittingValue = TRUE, ...)
}
\arguments{
\item{x}{The \code{DoseVolumeModel} class object}

\item{ShowFittingValue}{A logical value for showing the fitting point in the Log-Likelihood plot}

\item{...}{Other parameters to be passed to \code{plot} function}
}
\description{
This method plots the results of fitting process for a \emph{Dose/Volume} model. The \emph{x} axis represents either the Dose,
             or the Volume, the \emph{y} axis represents the Log-Likelihood result of the modeing process.
}
\details{
The \code{DoseVolumeModel} class object contains the results of a modeling process across different values of Dose or 
         Volumes (depending on which kind of fitting was chosen during fitting process). The plot doesn't show the impact of other covariates
         that could be part of the model itself.
}
