% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvh_mod.R
\docType{methods}
\name{plot,dvhmatrix,ANY-method}
\alias{plot,dvhmatrix,ANY-method}
\title{method for plotting \code{dvhmatrix} class objects}
\usage{
\S4method{plot}{dvhmatrix,ANY}(x, elements = NULL, enhance = NULL,
  mean.dvh = FALSE, median.dvh = FALSE, mean.median.alone = FALSE,
  el.color = "black", en.color = "red", mean.color = "blue",
  median.color = "red", lwd = 1, C.I.dvh = FALSE, C.I.dvh.width = 0.95,
  C.I.dvh.range = FALSE, n.boot = 2000, ...)
}
\arguments{
\item{x}{The \code{dvhmatrix} object to be plotted.}

\item{elements}{A vector containing the indeces of the DVHs to be plotted.}

\item{enhance}{A vector containing the indeces of the DVHs curves to be enhanced in the plot by changing de default color.}

\item{mean.dvh}{A \code{logical} value, if \code{TRUE} the mean dvh is plotted.}

\item{median.dvh}{A \code{logical} value, if \code{TRUE} the median dvh is plotted.}

\item{mean.median.alone}{A \code{logical} value, if \code{TRUE} only the mean and/or median DVH(s) is/are plotted.
Mean or median DVHs are plotted according the values in \code{mean.dvh} and \code{median.dvh}.}

\item{el.color}{The color for plotting DVHs. The default value is \code{"black"}.}

\item{en.color}{The color for plotting the enhanced DVHs. The default value is \code{"red"}.}

\item{mean.color}{The color for the mean dvh plot. The default value is \code{"blue"}.}

\item{median.color}{The color for the median dvh plot. The default value is \code{"red"}.}

\item{lwd}{The size of the lines in the plot. Default value is 1.}

\item{C.I.dvh}{Plot the confidence interval of mean and/or median DVH(s) according the values in \code{mean.dvh} and
\code{median.dvh}.}

\item{C.I.dvh.width}{The width of confidence interval of mean, median DVH(s) and overall dvh series.}

\item{C.I.dvh.range}{Plot the confidence interval of the whole DVHs series according the value in \code{C.I.dvh.width}.}

\item{n.boot}{The number of bootstrap repetitions for calculating the confidence interval of mean and median DVHs.}

\item{...}{Other parameters to be passed to \code{plot} function.}
}
\description{
This method overrides the default method \code{plot} and handles the plots of \code{dvhmatrix}
             class objects.
}
\details{
The dvh are stored in the slot \code{dvh} of a \code{dvhmatrix} class object. Of course it is possible for7
         users to use these matrices for plotting the dvhs considering that the structure of the matrix is referenced
         under \code{\link{dvhmatrix-class}}. In order to simplify the plotting operations this method has been coded
         allowing users to automatically obtain some useful features in the graphs, as the mean and median dvh curves,
         and the confidence intervals for mean, median and overall dvh matrix. The confidence intervals of mean and
         median are calculated by bootstrapping the whole \code{dvhmatrix} and calculating a given number of mean and
         median dvh values that finally are reduced by using a quantile function to create the final plot.
}
\examples{
## create a dvhmatrix object
b <- DVH.generate(dvh.number = 100, dvh.type = "cumulative", vol.distr = "absolute")
plot(x = b, mean.dvh = TRUE, median.dvh = TRUE, mean.median.alone = TRUE, 
     C.I.dvh = TRUE, C.I.dvh.range = TRUE, C.I.dvh.width = .67, lwd = 2)
}
