% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAD.features_00.R
\name{RAD.areaVolume}
\alias{RAD.areaVolume}
\title{function to calculate Are/Volume and related measures}
\usage{
RAD.areaVolume(listaROIVoxels)
}
\arguments{
\item{listaROIVoxels}{an output of a \code{obj$getROIVoxel()} method}
}
\value{
a list containing, for each patient the indicated measures
}
\description{
calculates Are, Volume, Area/Volume Ratio and equivolumetric Spherical Area Ratio
}
\examples{
\dontrun{
# Create an instante of mmButo and load some cases
obj<-mmButo()
obj$loadCollection(Path = '/progetti/immagini/urinaEasy')

# get the three ROIs
Retto<-obj$getROIVoxel(ROIName="Retto")  

# get the possible biopsy
uu<-RAD.areaVolume(listaROIVoxels = Retto)
}#' #' 
}
