% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAD.features_00.R
\name{RAD.ExtractFeatures}
\alias{RAD.ExtractFeatures}
\title{Extract a set of features}
\usage{
RAD.ExtractFeatures(whichFeatures = c(), obj.geoLet = obj,
  voxelCube = vc1)
}
\value{
many stuff
}
\description{
extract a set of features from a given voxel cube and geoLet obj
}
