% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAD.features_00.R
\name{RAD.firstOrderFeatureImage}
\alias{RAD.firstOrderFeatureImage}
\title{function to calculate the first order features}
\usage{
RAD.firstOrderFeatureImage(inputData)
}
\arguments{
\item{inputData}{a list where each element is an array of the voxel of the image. Each element of the list normally refers to a patient.}
}
\value{
six lists: the first list contains the entropies, the second the kurtosis, the third the skewness, the fourth the mean, the fifth the standard deviation and the sixth the energy
}
\description{
Calculates Shannon entropy, kursosis, Skewness, mean, standard deviation and energy of a given list of arrays of voxels
}
\examples{
\dontrun{
# Create an instante of mmButo and load some cases
obj<-mmButo()
obj$loadCollection(Path = '/progetti/immagini/urinaEasy')

# get the three ROIs
Retto<-obj$getROIVoxel(ROIName="Retto")  

# get the possible biopsy
aa<-RAD.firstOrderFeatureImage(inputData = Retto )
aa$entropy
}#' #' 
}
