% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAD.features_00.R
\name{RAD.getFeaturesListNames}
\alias{RAD.getFeaturesListNames}
\title{Return the list of the available features}
\usage{
RAD.getFeaturesListNames()
}
\value{
an array of strings
}
\description{
returns the list of the available features
}
