% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RADscout.r
\name{RAD.scoutV2}
\alias{RAD.scoutV2}
\title{class for loading and presenting DICOM data}
\usage{
RAD.scoutV2(path)
}
\description{
Instantiate an object of the class \code{geoLet}.This represents just the classname, 
              methods are exposed with the technique of 'closure'.
              In order to see manuals for the single mathods, consider the vignette or use the 
              available for the following wrapping functions:
              \itemize{
              \item \code{GLT.openDICOMFolder( );} : to load a DICOM series into an geoLet object
              }
              The original methods for the class geoLet can also be invocked using the same name without the previx 'GTL.', i.e.:
}
\examples{
\dontrun{
# first of all create an object aa
aa <- RAD.scoutV2(path = './RETTO/pazienti182/Pre/')


aa$scoutSigma(CSpatient = datasetALL$Codice.sanitario, outcome = datasetALL$TRG, ROIName = 'GTV', feature.family = c("stat"), interpolate = F, from.sigma = 0.3, to.sigma = 0.5, def.by = 0.1, forceRecalculus = F)
aa$pixelSpacing()




}
}
