% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_statisticalFeatures.R
\name{statisticalFeatures}
\alias{statisticalFeatures}
\title{class for loahvdv}
\usage{
statisticalFeatures(imgObj)
}
\description{
Instantiate an object of the class \code{geoLet}.This represents just the classname
}
